requirements.txt
google-genai
protobuf
python-dotenv
openai